Remove donload button and fix tooltip position
