+++ bigbluebutton-html5/imports/ui/components/button/component.jsx
@@ -115,6 +115,7 @@ export default class Button extends BaseButton {
       label,
       'aria-label': ariaLabel,
       'aria-expanded': ariaExpanded,
+      tooltipDistance,
     } = this.props;
 
     const renderFuncName = circle ? 'renderCircle' : 'renderDefault';
@@ -124,6 +125,7 @@ export default class Button extends BaseButton {
 
       return (
         <Tooltip
+          tooltipDistance={tooltipDistance}
           title={tooltipLabel}
         >
           {this[renderFuncName]()}
