+++ bigbluebutton-html5/imports/ui/components/presentation/download-presentation-button/styles.scss
@@ -7,7 +7,7 @@
 }
 
 .wrapper {
-  position: relative;
+  position: absolute;
   left: 0;
   background-color: var(--color-transparent);
   cursor: pointer;
@@ -36,4 +36,8 @@
 
 .dark {
   bottom: 0;
+}
+
+.dark .button span i {
+  color: var(--color-black);
 }
